pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "newsanalysis"
version = "2.0.0"
description = "AI-powered Swiss news analysis for credit risk intelligence"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Creditreform Switzerland"}
]

dependencies = [
    "openai>=1.47",
    "pydantic>=2.8",
    "pydantic-settings>=2.0",
    "trafilatura>=2.0.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.11.0",
    "lxml>=5.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "jinja2>=3.0.0",
    "structlog>=24.0.0",
    "click>=8.1.0",
    "sqlalchemy>=2.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "types-PyYAML",
    "types-requests",
    "types-python-dateutil",
]

playwright = [
    "playwright>=1.40",
]

cache = [
    "redis>=5.0",
    "sentence-transformers>=2.7",
    "faiss-cpu>=1.8",
]

[project.scripts]
newsanalysis = "newsanalysis.cli.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "feedparser",
    "trafilatura",
    "structlog",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=newsanalysis",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests that access external services",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
